recommendation_consumer/requirements.txt
kafka-python
mysqlclient==2.1.1 
SQLAlchemy==2.0.2
SQLAlchemy-Utils==0.40.0
